working on units import, fixed traceback
